pyproject.toml
[project]
name = "ballot-location"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "autoname>=2.0.1",
    "contextily>=1.6.2",
    "editdistpy>=0.1.5",
    "furl>=2.1.3",
    "geopandas>=1.1.2",
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "joblib>=1.4.2",
    "jupyter>=1.1.1",
    "keplergl>=0.3.7",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.5",
    "pylcs>=0.1.1",
    "pythainlp>=5.1.2",
    "shapely>=2.1.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "nbconvert>=7.16.6",
    "pre-commit>=4.0.0",
    "ruff>=0.11.13",
]
